[ modify ] DnDContext 抽出
